fix: throw when a requested field on CoMap.Record is missing
